Drop cluster-wide RBAC install and fix finalizer permissions issue.
We picked up an ArgoCD pattern where you can install with namespaced or
cluster-wide permissions. The cluster-wide was required so Argo could
potentially sync resources to the cluster it's running in. This however
is not relevant to the ApplicationSet CR, so we do not need to support
dual modes of install and can instead just deploy with namespaced RBAC.

Fix a missing permission for finalizers so we can set an Application
blockOwnerDeletion reference back to the ApplicationSet that is
originated from.